chore: align to comments

+++ apps/engine/priv/repo/migrations/20200326103841_create_blocks_table.exs
@@ -8,17 +8,17 @@ defmodule Engine.Repo.Migrations.CreateBlocksTable do
       # blocks state: forming, finalizing, pending_submission, submitted, confirmed
       add(:state, :string, null: false)
       # transaction order!
-      add(:nonce, :integer, null: false)
+      add(:nonce, :bigint, null: false)
       # plasma block number
-      add(:blknum, :integer, null: false)
+      add(:blknum, :bigint, null: false)
       # submitted transaction hash (gets updated with submitted_at_ethereum_height)
       add(:tx_hash, :binary)
       # at which height did we form the block
-      add(:formed_at_ethereum_height, :integer)
+      add(:formed_at_ethereum_height, :bigint)
       # doesn't mean mined! gets updated every time hash is submitted
-      add(:submitted_at_ethereum_height, :integer)
+      add(:submitted_at_ethereum_height, :bigint)
       # gas in wei
-      add(:gas, :integer)
+      add(:gas, :bigint)
       # mining is async and it might fail (like submitted with not enough gas, client error)
       add(:attempts_counter, :integer, default: 0, null: false)
       add(:inserted_at, :utc_datetime, null: false, default: fragment("now_utc()"))
